+++ config.sh
@@ -5,8 +5,6 @@
 #
 # Configuration file.
 
-. ~/.fancy-git/modules/config-manager.sh
-
 # include config-override.inc
 if [ -e ~/.fancy-git/config-override.sh ]; then
     . ~/.fancy-git/config-override.sh